use -Wall in kick and splines

+++ bitbots_dynamic_kick/src/kick_ik.cpp
@@ -22,7 +22,7 @@ void KickIK::reset() {
    * first step will be not correct */
   std::vector<std::string> names_vec = {"LHipPitch", "LKnee", "LAnklePitch", "RHipPitch", "RKnee", "RAnklePitch"};
   std::vector<double> pos_vec = {0.7, 1.0, -0.4, -0.7, -1.0, 0.4};
-  for (int i = 0; i < names_vec.size(); ++i) {
+  for (size_t i = 0; i < names_vec.size(); ++i) {
     goal_state_->setJointPositions(names_vec[i], &pos_vec[i]);
   }
 }
